Return player to the hub if they log in an island world while not having an island.
This would happen if, for example, the faction was deleted (disbanded) while the player logged out on the faction island. The faction island would be deleted and they would fall into the void. I pray that PlayerJoinEvent triggers first so the player does not wait for their faction / island to be fetched.