Fix: Do not add obelisk entity if external chunk.
On platforms such as MultiPaper, where world(s) are split over multiple server instances, we should only store instances of an obelisk block entity if the location of the entity belongs to the server instance the plugin is running on.

We make this simple call (`MultiLib#isChunkExternal(Entity)`)  and return if it is true, meaning the entity does not belong to the current server instance.

This is still backwards compatible with all other platforms thanks to the devs of MultiLib.